feat: Update Footer component with new text and styling
This commit updates the Footer component by replacing the existing text with a new message in Spanish, "Gracias por tu interés | Desarrollado por Yessenia Villarte". It also adds a new CSS class, "dark:text-white", to change the text color to white when using a dark theme. This change enhances the content and visual appearance of the footer section.